design related refactoring
